Revert wrong change in nikita ppo

+++ mushroom_rl/algorithms/actor_critic/deep_actor_critic/ppo_nikita.py
@@ -36,7 +36,7 @@ class NikitaPPO(PPO):
 
         old_pol_dist = self.policy.distribution_t(state_old)
 
-        old_log_p = old_pol_dist.log_prob(action.to(dtype=torch.float32))[:, None].detach()
+        old_log_p = old_pol_dist.log_prob(action)[:, None].detach()
 
         self._V.fit(state, v_target, **self._critic_fit_params)
